requirements.txt
Flask
vinted-api-wrapper